requirements.txt
scikit-learn
rknn-toolkit2
onnx

